cleanup header files
