+++ includes/lexer_prototypes.h
@@ -1,0 +1,29 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   lexer_prototypes.h                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cwenz <cwenz@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/12 14:36:55 by cwenz             #+#    #+#             */
+/*   Updated: 2023/11/12 14:55:02 by cwenz            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef LEXER_PROTOTYPES_H
+# define LEXER_PROTOTYPES_H
+
+#include "structs.h"
+
+t_lexer	*tokenise(char *input);
+int		token_word(t_lexer **head, char *str);
+int		token_pipe(t_lexer **head, char *str);
+int		token_squote(t_lexer **head, char *str);
+int		token_dquote(t_lexer **head, char *str);
+int		token_backslash(t_lexer **head, char *str);
+int		create_less(t_lexer **head, char *str);
+int		create_greater(t_lexer **head, char *str);
+int		create_dbless(t_lexer **head, char *str);
+int		create_dbgreater(t_lexer **head, char *str);
+
+#endif
