+++ includes/parser_prototypes.h
@@ -1,0 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   parser_prototypes.h                                :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cwenz <cwenz@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/12 14:47:43 by cwenz             #+#    #+#             */
+/*   Updated: 2023/11/12 14:54:58 by cwenz            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef PARSER_PROTOTYPES_H
+# define PARSER_PROTOTYPES_H
+
+# include "structs.h"
+
+int		parser(t_lexer **lexer, t_minishell *minishell);
+int		count_pipes(t_lexer **lexer);
+void	init_t_cmd(t_cmd **cmd);
+void	free_cmd_table(t_cmd **cmd_table);
+int		count_words(t_lexer **lexer);
+char	*expander(t_lexer *lexer, t_minishell *minishell);
+void	print_cmd_table(t_cmd **cmd_table);
+int		syntax_checker(t_lexer **lexer, t_minishell *minishell);
+
+#endif
