+++ includes/signal_prototypes.h
@@ -1,0 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   signal_prototypes.h                                :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cwenz <cwenz@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/12 14:50:09 by cwenz             #+#    #+#             */
+/*   Updated: 2023/11/12 14:54:54 by cwenz            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef SIGNAL_PROTOTYPES_H
+# define SIGNAL_PROTOTYPES_H
+
+void	setup_signals();
+
+#endif
